Missing header, move const to AuthUtils
