+++ frontend/src/components/Layout/Header.tsx
@@ -45,7 +45,7 @@ const StandaloneHeader: FC = () => {
 const MultiUserHeader: FC<{ handleDrawerOpen: () => void }> = ({
   handleDrawerOpen,
 }) => {
-  const showTabsRegex = /^\/dashboard\/workspaces\/.+$/
+  const showTabsRegex = /^\/workspaces\/.+$/
   const location = useLocation()
 
   return (
